+++ samples/Calculator/CalculatorServiceManager.cs
@@ -6,7 +6,7 @@ namespace Calculator
     public class CalculatorServiceManager : ServiceManager
     {
         public CalculatorServiceManager()
-            : base("4.0", new DefaultSchemaExporter("http://calculator.x-road.eu/", typeof(CalculatorServiceManager).Assembly))
+            : base("4.0", new DefaultSchemaProvider("http://calculator.x-road.eu/", typeof(CalculatorServiceManager).Assembly))
         { }
     }
 }