+++ test/XRoadLib.Tests/Globals.cs
@@ -7,7 +7,7 @@ namespace XRoadLib.Tests
 {
     public static class Globals
     {
-        public static readonly ServiceManager ServiceManager = new("4.0", new DefaultSchemaExporter("http://test-producer.x-road.eu/", typeof(Class1).Assembly) { SupportedVersions = { 1u, 2u, 3u } });
+        public static readonly ServiceManager ServiceManager = new("4.0", new DefaultSchemaProvider("http://test-producer.x-road.eu/", typeof(Class1).Assembly) { SupportedVersions = { 1u, 2u, 3u } });
         public static readonly DirectoryInfo StoragePath = new(Path.GetTempPath());
 
         public static ContentDefinition GetTestDefinition(Type type)